+++ presto-main/src/main/java/com/facebook/presto/event/query/QueryCompletionEvent.java
@@ -108,7 +108,7 @@ public class QueryCompletionEvent
         if (duration == null) {
             return null;
         }
-        return (long) duration.toMillis();
+        return duration.toMillis();
     }
 
     @Nullable
